pw-cli: Collect and display parameters from global symbols
Until now there was no method to display parameters along with
information of their parent object.

This patchset stores parameter info in the global object so
that it may displayed along with its parent object.

This is a necessary step to be able to display and modify parameters
in a simple form like the following

pipewire-0>> set 12 volume=0.7
pipewire-0>> get 12 volume
0.7

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>