node: use intermediate pending state
Use an intermediate pending state that contains the state currently
in progress and use this to decide if we need to issue start/stop
commands.

Otherwise we might be in the running state, issue a Stop command,
issue a Start command (ignored because already running), then the
Stop completes and the node is stopped.